+++ packages/sit-onyx/src/components/OnyxTab/OnyxTab.stories.ts
@@ -22,3 +22,17 @@ export const Default = {
     default: "Panel content 1...",
   },
 } satisfies Story;
+
+export const Disabled = {
+  args: {
+    ...Default.args,
+    disabled: true,
+  },
+} satisfies Story;
+
+export const Skeleton = {
+  args: {
+    ...Default.args,
+    skeleton: true,
+  },
+} satisfies Story;